New branch

+++ server/src/main/java/server/jpa/WorkerRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface WorkerRepository extends CrudRepository<Worker, Long> {
     List<Worker> findWorkerByLogin(String login);
     List<Worker> findWorkerByLoginAndPassword(String login, String password);
-    List<Worker> findWorkersByIdIn(List<Integer> ids);
+    List<Worker> findWorkersByIdIn(List<Long> ids);
     @Query(value = "SELECT * FROM WORKERS WHERE CAST((start_date - CURRENT_DATE) AS INT)%30=0", nativeQuery = true)
     List<Worker> findToPayment();
 }
